page frame allocator, page table manager, recursive paging

+++ src/initrd.h
@@ -3,7 +3,7 @@
 
 struct initrd {
     
-    uint32_t address;
+    uint64_t address;
     uint32_t size;
     
 }__attribute__((packed));
